requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0
sqlalchemy==1.4.50
alembic==1.12.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Platform APIs
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2

# NLP & Analysis
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0